+++ coffeenet-autoconfigure/src/main/java/coffee/synyx/autoconfigure/discovery/config/CoffeeNetDiscoveryClientProperties.java
@@ -1,0 +1,16 @@
+package coffee.synyx.autoconfigure.discovery.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import org.springframework.cloud.netflix.eureka.EurekaClientConfigBean;
+
+import org.springframework.validation.annotation.Validated;
+
+
+/**
+ * @author  Tobias Schneider
+ */
+@Validated
+@ConfigurationProperties(prefix = "coffeenet.discovery.client")
+class CoffeeNetDiscoveryClientProperties extends EurekaClientConfigBean {
+}